otel update
